Replace winapi GetQueuedCompletionStatus bind with x/sys/windows (#1307)
Previously we had our own definition for GetQueuedCompletionStatus as x/sys/windows
had an incorrect definition for it. This was remedied a bit ago in this change
https://github.com/golang/sys/commit/683adc9d29d7bd1f0f778e1f7f2eeb61b415f4d5
so we're alright to remove our own at this point.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit 47214119f46d2e745b48f5f03400da710d432e42)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>