GCS changes to detect SCSI devices.
This commit modifies the GCS code to look for new SCSI devices attached to the system
instead of relying on the controller & LUN numbers provided by hcsshim. We add new
wrappers on top of the Mount, Unmount and UnplugDevice SCSI functions to identify the
correct controller LUN numbers first and then call into the actual functions.

Signed-off-by: Amit Barve <ambarve@microsoft.com>