Go 1.17 is the minimum version in all cases
This ensures:
- CI explicitly requires Go 1.17
- README specifies Go 1.17
- Tests also depend on Go 1.17

The go.mod already specifies 1.17 as the minimum Go version, so this
shouldn't be a difference in practice.

GitHub Actions setup-go was installing Go 1.17 already, presumably
honouring the value in go.mod.

One interesting change is that in 1.17, go.mod lists indirect
dependencies, and vendoring does not capture go.mod/go.sum, so the
test/vendor directory got slightly less noisy.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>