Adding build constraints
Adding windows build constraints to code to allow tests and benchmarks
to be run on Linux.

Added doc.go to modules (with doc string, where appropriate) to prevent
compiler/linter errors about broken imports.
In some cases (ie, winapi and wclayer), the package already had an
OS-agnostic file of the same name, along with a doc string. A doc.go
file was added to preempt situations where windows-specific code is
added to that file in the future.

Signed-off-by: Hamza El-Saawy <hamzaelsaawy@microsoft.com>