Add Go bindflt/silo definitions (#1331)
* Add Go bindflt/silo definitions

This change adds a couple Go bindings for calls and constants from bindfltapi.dll
as well as some silo flags for job objects. Together these allow file bindings
(think bind mounts on Linux) to be performed for a specific job object. The bindings
are only viewable from processes within that specific job.

This additionally adds a couple unit tests for this behavior.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>