Implement file binding support for host process containers
This change adds in file binding support for host process containers
if the host has the functionality available (bindfltapi.dll exists).
This makes it so mounts in the runtime spec actually show up in the
container at mount.Destination instead of simply being symlinks to a
relative path, as well as are completely unique per container. This is
achieved by upgrading the job object to a silo and making use of silo
local file bindings that the Bind Filter supports.

This support additionally opens up the opportunity for the rootfs for the
container to be container local and unique. So instead of the rootfs showing
up to any process on the host and being located at C:\C\<containerid>, it
now (by default at least) will be present at C:\payload and be unique in
each container as well. In a similar fashion to the mount changes, this
only takes affect if the host has file binding support available.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>