Pin go version for linter to 1.17.x
Some of the tooling golangci-lint uses doesn't fully support 1.18 yet
resulting in a bunch of hard to decode errors. We were using ^1.17.0
as our version listed so we ended up resolving to 1.18 a couple of days
ago and finally ran into this. v1.45.0 of golangci-lint has a
workaround for this which is disabling some of the problematic linters,
but these are some of our most used. This seems like a sane fix for
now until the kinks are worked out and things are working on 1.18.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>