Support for LCOW tests and benchmarks
Updated go-winio to include bugfixes for closing hvsockets, specifically
for writing (which is needed by `internal\cmd` to signal that the stdin
stream has finished.

Expose or add certain functionality needed by tests and benchmarks both
in the functional directory and uploaded directly onto a Linux UVM.
For example, ability to access an hcsv2 containers init process and ID,
as well creating and deleting network namespaces on the Linux UVM.

Added more functionality to `internal\tools\uvmboot`, to include adding
SCSI VHD and file mounts, as well as disabling the time sync, and setting
the uVM security policy.

Added winapi call to check if process is elevated, to allow returning a
clear error message before attempting to start uVMs.

Split out `internal\guest\runtime\runc\runc.go` into `runc.go`,
`container.go` and `process.go`.

Reorganized makefile to read from top to bottom, and added additional
files to LCOW rootfs that include the time stamp of of the vhd creation,
the image build date, and its full name (pulled from a
`*.testdata.json` file in the LSG release, that appears to be
the only location of that information).

Signed-off-by: Hamza El-Saawy <hamzaelsaawy@microsoft.com>