Optimize Windows container statistics
This change adds in a optimization workaround for cases where we query
HCs for just statistics through our internal PropertiesV2 interface.

The optimization we can make here if the user is asking for
Statistics only, and this is due to an inefficient way that HCS calculates
the private working set total for a given container. HCS calls
NtQuerySystemInformation with the class SystemProcessInformation which
returns an array containing system information for *every* process
running on the machine. They then grab the pids that are running in
the container and filter down the entries in the array to only what's
running in that silo and start tallying up the total. This doesn't work
well as performance should get worse if more processess are running on the
machine in general and not just in the container. All of the additional
information besides the WorkingSetPrivateSize field is ignored as well
which isn't great and is wasted work to fetch.

HCS only let's you grab statistics in an all or nothing fashion, so we
can't just grab the private working set ourselves and ask for everything
else seperately. The optimization we can make here is to open the silo
ourselves and do the same queries for the rest of the info, as well as
calculating the private working set in a more efficient manner by:

1. Find the pids running in the silo
2. Get a process handle for every process (only need
PROCESS_QUERY_LIMITED_INFORMATION access)
3. Call NtQueryInformationProcess on each process with the class
ProcessVmCounters
4. Tally up the total using the field PrivateWorkingSetSize in
VM_COUNTERS_EX2.

This change additionally:
* Changes the jobcontainers package to use this new way to calculate the
private working set.
* Change the query the StorageStats method in the jobobject package uses
to grab IO counters to match what HCS queries.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>