Add secure hardware support for uvmboot
Fix a bug where resource modifications are not rejected when running
in non-snp mode by setting a default open door policy when no
security policy is passed and making sure that security policy is
always set for UVM. Additionally, don't assume that security hardware
is present when a security policy is set explicitly.

Add a flag to boot from VMGS file.

Signed-off-by: Maksim An <maksiman@microsoft.com>