Fix Hyper-V check in late clone spec comparisons (#1400)
In the function that handles checking if the cloned containers spec
retrieved from the registry matches the current spec, it was comparing
the hyperv field on the runtime spec using the != operator which won't
work as the hyperv field is a pointer so it would just be comparing the
addr. Swap to reflect.DeepEqual.

This became an 'issue' as we set the hyperv field now if in our shim
options SandboxIsolation is set to the HYPERVISOR option. This doesn't
have much of an effect being set for containers that are going to be launched
IN the UVM, so this is mostly just a bug that was surfaced from a field
that didn't use to be set.

This additionally changes to errors.New instead of fmt.Errorf where
there was no formatting in the error.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>