Add annotations passthrough for host process containers
This changes adds in a way for annotations specified for a sandbox
container to be passed through to every container in the pod.
K8s only passes annotations to the RunPodSandbox call and not to individual
containers unfortunately. To accomplish this I cache the pod sandboxes OCI
spec on the pod object as well as expose a method that passes through
specific annotations from the pod spec through to an individual containers.

This is useful for a couple of host process containers annotations, like
microsoft.com/hostprocess-rootfs-location which specifies a non-default
path for the rootfs of the container to show up at.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>