Safeguards on slicing hostprocess username
Adds some checks to make sure the container ID we get from the runtime
is greater than the max character limit for a username before slicing
it down.

HostProcess containers are a k8s feature for the most part and
Containerd's container ID generation for CRI outputs IDs that are
> 20 characters, however it's not good to rely on this implementation
detail. Other containerd clients (ctr, nerdctl) trying to launch a
hostprocess container manually with a provided ID would panic here
if it was < 20 characters.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>