Fix up hostprocess integration tests
This fixes a couple issues with the host process integration tests.

- Instead of passing the annotation to ask for a host process pod, use the
hostprocess CRI fields instead. The annotations used to work, however
containerd now overrides these with the value of the hostprocess CRI field
so this is the only way to ask for hpc now.

- The VHD test would hit a "Paramter is incorrect" error on attaching the vhd.
Not sure what's causing this but it's probably the set of flags used to open
the vhd inside AttachVhd, as supplying no flags has the open succeed.

- Now that there's new behavior for mounts for hpc if the bindflt dll is
available, gate some of the tests that tested the old behavior behind
rs5 for now.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>