Correctly set silo field when opening job object
We don't set the silo field on Open of an existing job object today.
This is useful if once opening a job we want to bind a file that only
that silo can see as it relies on atomically checking the `silo` u32
field to determine if we can carry out the operation.

The manner in which we check if the job is a silo is by using a new
jobobject information class with QueryInformationJobObject that fails
unless the job is a silo.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>