Remove log file from runc commands
Certain runc commands (eg, Delete, State) do not need to use a log file
since they do not pass stdio to the container.
This PR switches the commands to consume the error directly from stderr,
without needing to parse from a file.

This fixes a bug where the directory containing the runc log file for a
container is deleted before the container itself can be deleted via the
runc command.

Signed-off-by: Hamza El-Saawy <hamzaelsaawy@microsoft.com>