Add CI stage dependencies
Reorganize CI to run in stages, so lint and protobuild are dependencies
for vendoring, which in turn is required for testing then building.

This allows issues to be caught early on, and prevents long-running
test stages to run before lint or vendor issues fail CI.

Ran YML formatter on ci.yml file.

Added `if` condition to uploading test binary artifices for non PR
runs, to avoid unnecessary uploading and storage.

Signed-off-by: Hamza El-Saawy <hamzaelsaawy@microsoft.com>