Rework /internal/queue package (#1449)
* Rework /internal/queue package

Given our use cases for this package, we don't need methods that don't block
on reads if there's no value to be read. Due to this, I've removed the
ReadOrWait function and did a small redesign of the methods to be more
in line with standard queue method naming.

* Change Read/Write/IsEmpty to Dequeue/Enqueue/Size and remove ReadOrWait.
Now there is no version of Read/Dequeue that doesn't block if the queue
is empty.
* Fix up tests to be in line with this removal of the non-blocking read
and simplified most of the tests.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit 12d4cd8f9abaf895ee4a2daa6c5837554b92e33d)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>