Add IO tracking option for job objects
HCS enables this to get more in depth IO stats for the silo of the
container. I'd swapped hostprocess containers to querying for
these stats but without the prerequisite of actually enabling them :)

This change adds a new option on jobobject.Options{} to enable this
functionality and adds a new test to ensure we can actually call
StorageStats now.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>