Backwards compat for hostprocess cntrs mounts
In beta, mounts couldn't be unique per silo so they were mounted
to a relative path under the rootfs for the container. There's a
bunch of pod specs/apps floating around that were written with that
in mind so as to make a smooth transition, additionally keep mounting
them under the rootfs for now via the same approach (symlink the dir/file).

Signed-off-by: Daniel Canter <dcanter@microsoft.com>