Readonly option for hostprocess mounts
Add in ability to parse read only mount options for hostprocess mounts.
Change the API of ApplyFileBinding to take in a readOnly bool instead of
it signifying a merged binding. We don't have any use for merged bindings
as the default for containers is to shadow the directory we're binding
to. We could alternatively pass in a set of options/flags as this argument,
but readonly seems to be the only thing needed as of now.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>