Add shim command for launching hostprocess cntr processes
Due to the rootfs for a hostprocess container being a bind mount that is
only viewable for a process currently running in the containers silo, this
poses some challenges for launching a process in the container image via
CreateProcess from outside of the container. There's no usermode mechanism
to join a job object/silo and then leave afterwards, so that isn't viable either.

How we were working around this fact today is by launching container processes
via cmd /c C:\hpc\mybinary.exe rest of the commandline. The reason this
works is because cmd exists outside of the bind mount and should always
exist, and then once cmd starts up it is able to view the bind mount as
it's running in the containers silo. We can get the same functionality
and keep full control of launching and managing the process by exposing a
new command on our shim that we can invoke instead.

This change adds in a new hidden hpc command to our shim that host process
containers will make use of. It briefly communicates with the parent shim
instance over a named pipe to grab the process launch info and tell the
parent if anything prior to launching the process went wrong. This is a big
win as with using cmd /c didn't afford us this, we'd only learn if the process
couldn't be found/started from stdio of the cmd instance, so kubernetes wouldn't
count this as a startFailure for instance.

This new shim instance will join the containers silo so it can successfully
see the rootfs (generally at C:\hpc) and launch processes. This also allows
us to get some backwards compatible behavior for how bind mounts are handled
on systems that bindflt isn't available as well. In addition to binding mounts to
the path requested, we can additionally symlink them to a relative path under
the containers rootfs, which is what was done for non-bindflt machines. We
were currently working around this by temporarily mounting the volume
under C:\temp, setting up the symlinks, and then unmounting.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>