add `-snp` flag to gcs binary
This PR adds a way to tell GCS that it is expected to be running
inside an SEV-SNP enabled linux.
In addition to the flag we also try making "get-snp-report" ioctl
and refuse to run when `-snp` is requested, but not supported or
when `-snp` is not requested, but supported.

Signed-off-by: Maksim An <maksiman@microsoft.com>