Revert "Pass disk handle for computestorage.FormatWritableLayerVhd on RS5 (#1204)"
This reverts commit aaf5db90ef6961e767a4d5ca4bcf7f1f6465bcca.

We'd added a change to FormatWritableLayerVhd to help the caller work around
a breaking change in the OS, but this would actually cause a breaking change
in our wrapper of it if the caller was already working around the issue
themselves. To avoid this scenario, revert the commit that added the
"friendly" behavior.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>