Windows driver installation tool
Added functionality to `device-util` tool to install legacy-type
drivers; and list services and drivers.
`device-util` was updated to use `urface/cli/v2`. The original
read object and query children commands were kept untouched, although
aliases were added to be consistent with the driver and service sub
commands.

This is for drivers that cannot be installed via `pnputils`, which
instead are installed via creating a `kernel-driver` type service.

Since uVM WCOW base images (nanoserver) do no come with `sc.exe`, which
is what would normally be used in this situation, this code replicates
a small portion of its functionality, but is not intended to be a
drop-in replacement, or remotely as fully featured.

Signed-off-by: Hamza El-Saawy <hamzaelsaawy@microsoft.com>