Stub out Rego policy enforcer and hide it behind a build tag.
Add enforcer registration logic and support for default enforcer.
The host can request which security policy enforcer to use with
supplied policy, if none supplied, GCS code tries to make a "guess"
as to which enforcer should be used: "allow all" or "default".
Default enforcer is set to `StandardSecurityPolicyEnforcer` unless
GCS is built with "rego" tag present. In that case, the default
enforcer will be set to `RegoEnforcer`.

New annotation has been added that allows callers to pick which
enforcer to use, e.g.
```pod.json
{
    ...
    "annotations": {
        "io.microsoft.virtualmachine.lcow.enforcer": "rego"
    },
    ...
}
```

Signed-off-by: Maksim An <maksiman@microsoft.com>