Remove wait mounts functionality (#1474)
Wait mounts was originally added as a synchronization mechanic that
allows to build dependencies between containers via security policy.

Some issues:

- Only available via policy so, you can't use in a non-confidential scenario
- Mixes configuration into policy enforcement
- Completely bespoke

With our moving to a rego based policy engine, we've decided to stop doing
anything that is "include configuration in policy" and instead switch to
"policy controls if a bit of configuration is allowed to be changed". This
approach eliminates wait mounts as a possibility to move over even if we
decided that the issues listed above weren't problematic.

We expect that any customer using wait mounts functionality via policy
will switch to handling dependencies amongst containers as they usually
would. Allow the container that is missing a dependency to fail in a fashion
that will get the orchestrator to restart it until such time as the dependency
is available. For wait mounts, this means exiting if the mounted drive is
missing a known good file that is expected to exist thereby indicating that
the drive in question wasn't available at the time the container started and
instead, an empty directory was mounted.

These changes are a prerequisite for the soon to arrive change to switch from
the bespoke JSON based policy language I wrote to using Rego.

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>