Add Tasks command to shimdiag
This command prints the tasks that are currently being managed
by a given shim instance, and optionally the execs in each task.

There've been a couple times where our shim and containerd have
seemingly disagreed on what tasks are still running.. This should
make it easy to see the state of the world from our shims point
of view.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>