Fix nil pointer dereference in addSCSI
The new change that we added to fix a race condition in addSCSI introduced a bug where the code ends up accessing a
nil pointer in certain situations. For example, the deferred function to unblock any waiters of the attach
SCSI operation accesses the scsi mount object to propagate any errors. However, this pointer is a named return
value of the function and is set to `nil` when returning an error. In those cases the deferred function panics
with the nil pointer dereference error. To fix this we don't use the named return value for the scsi mount
object anymore.
This change also removes the check for zero SCSI controllers since that check is done by the
`allocateSCSIMount` function.

Signed-off-by: Amit Barve <ambarve@microsoft.com>