Fix seeding for generative policy tests (#1508)
We are seeding the `testRand` module level variable. However, testing/quick
was not set to use that seeded random number generator. The testing/quick
configuration has a field `Rand` which is not set, will result in a
random number generator being created and used. This created generator
hasn't been seeded so our seeding had no impact on any of our `Generate`
functions.

This commit updates all testing/quick tests to initialize `Rand` in the
configuration to be set to our seeded `testRand`.

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>