Adding the exec_in_container enforcement point
This commit adds the exec_in_container enforcement point, but as part of this change we must also perform a
refactoring of the framework to enable supporting this and other additional enforcement points which come
after creating a container. Due to the number of possible matches and the way in which each subsequent
enforcement point narrows the list of matches, we need to maintain for each container ID a list of potentially
matching containers.

As part of making this change to the framework, we took the opportunity to build a more flexible system for a
policy to store data for use in evaluating later rules. This has the beneficial side effect of removing more policy
logic from the Go code while also creating a far more powerful and flexible system for policy authoring.

Before, a policy would define a simple true/false rule for something like mount_device. Now, the mount_device rule
(along with all others) is expected to return an object, as seen below:

```rego
device_mounted(target) {
    data.metadata.devices[target]
}

default mount_device := {"allowed": false}

mount_device := {"devices": devices, "allowed": true} {
    not device_mounted(input.target)
    some container in data.policy.containers
    some layer in container.layers
    input.deviceHash == layer
    devices := {
        "action": "add",
        "key": input.target,
        "value": input.deviceHash
    }
}
This object contains a member called allowed which indicates whether the operation should proceed, but also includes one or more "metadata" commands, of the following form:

{
    "<name>": {
        "action": "<add|update|remove>",
        "key": "<key>",
        "value": "<optional value>"
    }
}
```

These metadata commands alter a special metadata namespace. The Go code, which previously contained logic for maintaining
various data structures for use by the framework, now executes these metadata commands instead. This both means that the
Go code contains almost no policy logic at this point, but also that authored policies can take advantage of all the same
kinds of data caching logic upon which the framework is based.

A consequence of this change is that unmount_device is now a new enforcement point, allowing policy authors to have control
from the Rego side of how devices are unmounted (as they do with mounting devices at the moment).

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>